+++ app/api/entities/unit_entity.rb
@@ -39,6 +39,8 @@ module Entities
     end
 
     expose :active
+    expose :grade_values
+    expose :grade_definitions
 
     expose :overseer_image_id, unless: :summary_only, if: lambda { |unit, options| can_read_unit_config?(options[:my_role]) }
     expose :assessment_enabled, unless: :summary_only