+++ app/helpers/authentication_helpers.rb
@@ -221,7 +221,7 @@ module AuthenticationHelpers
 
       # Generate a new token when the old one is absent or getting close to expiring
       if token.nil? || token.auth_token_expiry <= Time.zone.now - 12.hours
-        token = current_user.generate_authentication_token!(token_type: :refresh_token, expiry: Time.zone.now + 1.week)
+        token = current_user.generate_authentication_token!(token_type: :refresh_token)
       end
 
       domain = Doubtfire::Application.config.institution[:cookie_domain]