+++ app/models/pdf_generation/project_compile_portfolio_module.rb
@@ -71,7 +71,9 @@ module PdfGeneration
         @doubtfire_product_name = Doubtfire::Application.config.institution[:product_name]
         @is_retry = is_retry
         @include_pax = !is_retry
-        @work_id = "portfolio-#{project.id}-#{Time.now.to_i}-#{Process.pid}-#{Thread.current.object_id}#{'-retry' if is_retry}"
+        @work_id = FileHelper.sanitized_path(
+          "portfolio-#{Time.current.strftime('%Y%m%d-%H%M')}-#{project.student.username}-#{project.id}-#{Process.pid}#{'-retry' if is_retry}"
+        )
       end
 
       def make_pdf
@@ -113,9 +115,7 @@ module PdfGeneration
 
     # Create the portfolio for this project
     def create_portfolio
-      self.compile_portfolio = false
-      save!
-
+      logger.info "Creating portfolio for #{user.username} in #{unit.code}"
       begin
         pac = ProjectAppController.new
         pac.init(self, false)
@@ -142,8 +142,13 @@ module PdfGeneration
         logger.info "Created portfolio at #{portfolio_path} - #{log_details}"
 
         self.portfolio_production_date = Time.zone.now
-        save
+        self.compile_portfolio = false
+        save!
+        true
       rescue StandardError => e
+        self.compile_portfolio = false
+        save!
+
         logger.error "Failed to convert portfolio to PDF - #{log_details} -\nError: #{e.message}"
 
         log_file = e.message.scan(%r{/.*\.log}).first
@@ -197,6 +202,19 @@ module PdfGeneration
           task.task_definition.upload_requirements.blank? &&
           ![TaskStatus.need_help.id, TaskStatus.working_on_it.id].include?(task.task_status_id)
         )
+      end
+    end
+
+    # Return the tasks that are currently being processed
+    def tasks_processing_pdf
+      # Get assigned tasks that should be included in the portfolio
+      tasks = self.tasks.joins(:task_definition).order('task_definitions.target_date, task_definitions.abbreviation')
+
+      # Select tasks that should have a PDF submission, but is currently being processed
+      tasks.select do |task|
+        !task.has_pdf &&
+          task.processing_pdf? &&
+          task.task_definition.upload_requirements.present?
       end
     end