+++ test/factories/submission_histories.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :submission_history do
+    task
+    sequence(:submission_timestamp) { |n| "#{Time.current.to_i}-#{n}" }
+  end
+end